+++ R/helpers_anova.R
@@ -603,22 +603,10 @@ expr_anova_robust <- function(data,
 #' set.seed(123)
 #' library(statsExpressions)
 #'
-#' # between-subjects ---------------------------------------
-#'
 #' expr_anova_bayes(
 #'   data = ggplot2::msleep,
 #'   x = vore,
 #'   y = sleep_rem
-#' )
-#'
-#' # repeated measures ---------------------------------------
-#'
-#' expr_anova_bayes(
-#'   data = WRS2::WineTasting,
-#'   x = Wine,
-#'   y = Taste,
-#'   paired = TRUE,
-#'   k = 4
 #' )
 #' @export