+++ NEWS.md
@@ -1,4 +1,4 @@
-# statsExpressions 0.6.0.9000
+# statsExpressions 0.6.0
 
 BREAKING CHANGES